refactor handler, move handler/index to _run_server
